+++ demo/ortho_test.cpp
@@ -54,12 +54,12 @@ int main() {
 
     scene.setCamera(camera);
 
-    int totalTime = 0;
+    float totalTime = 0.0f;
 
-    scene.setUpdateCallback([&](knot::Scene& currentScene, int deltaTime) {
+    scene.setUpdateCallback([&](knot::Scene& currentScene, float deltaTime) {
         totalTime += deltaTime;
 
-        cubeObject->rotation = glm::quat(glm::vec3(sin(static_cast<float>(totalTime) * 0.0005f) * 0.2f, static_cast<float>(totalTime) * 0.0005f, 0.0f));
+        cubeObject->rotation = glm::quat(glm::vec3(sin(totalTime * 0.0005f) * 0.2f, totalTime * 0.0005f, 0.0f));
     });
 
     engine.setScene(scene);
